+++ sample/src/main/java/id/co/veritrans/mdk/v1/sample/util/SpringDataInitializerBean.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
-import java.math.BigDecimal;
 import java.util.Arrays;
 
 /**
@@ -22,12 +21,12 @@ public class SpringDataInitializerBean {
     public void setup() {
         final Product p1 = new Product();
         p1.setShortName("Apple");
-        p1.setPriceIdr(new BigDecimal(10000));
+        p1.setPriceIdr(new Long(10000));
         p1.setThumbnailFilePath("images/products/apple.jpg");
 
         final Product p2 = new Product();
         p2.setShortName("Orange");
-        p2.setPriceIdr(new BigDecimal(20000));
+        p2.setPriceIdr(new Long(20000));
         p2.setThumbnailFilePath("images/products/orange.jpg");
 
         productRepo.save(Arrays.asList(p1, p2));
